requirements.txt
python-binance
httpx